Section_12_Bonus/Aula_1_charts/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.5
django-bootstrap4==3.0.1
django-chartjs==2.2.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
